+++ src/adjusters/number.mjs
@@ -4,7 +4,8 @@ import Default from "../decorators/default";
 import AcceptNull from "../decorators/acceptNull";
 import AcceptEmptyString from "../decorators/acceptEmptyString";
 import Only from "../decorators/only";
-import Map from "../decorators/map";
+import Convert from "../decorators/convert";
+import AcceptFullWidth from "../decorators/number/acceptFullWidth";
 import Type from "../decorators/number/type";
 import MinValue from "../decorators/number/minValue";
 import MaxValue from "../decorators/number/maxValue";
@@ -21,11 +22,12 @@ export default () =>
 /**
  * adjuster for number
  */
-@Map
+@Convert
 @MaxValue
 @MinValue
 @Only
 @Type
+@AcceptFullWidth
 @AcceptEmptyString
 @AcceptNull
 @Default