+++ src/libs/constants.mjs
@@ -9,7 +9,8 @@ export const CAUSE = {
 	NULL: "null",
 	EMPTY: "empty",
 	ONLY: "only",
-	MAP: "map",
+	MAP: "convert", // deprecated; use CONVERT
+	CONVERT: "convert",
 
 	MIN_VALUE: "min-value",
 	MAX_VALUE: "max-value",