+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.4.0] - 2019-04-15
+
+### Added
+
+* `adjuster.number().acceptFullWidth()`
+* `adjuster.numericString().fullWidthToHalf()`
+* `adjuster.number().convert()`
+* `adjuster.string().convert()`
+* `adjuster.numericString().convert()`
+
+### Deprecated
+
+* `adjuster.number().map()`
+* `adjuster.string().map()`
+* `adjuster.numericString().map()`
+
 ## [1.3.1] - 2019-03-09
 
 ### Added
@@ -338,7 +354,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * First release.
 
-[Unreleased]: https://github.com/shimataro/node-adjuster/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/shimataro/node-adjuster/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/shimataro/node-adjuster/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/shimataro/node-adjuster/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/shimataro/node-adjuster/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/shimataro/node-adjuster/compare/v1.1.0...v1.2.0